requirements.txt
Django==1.9.1
django-crispy-forms==1.6.0
django-filter==0.12.0
djangorestframework==3.4.4
Markdown==2.6.5
Pillow==3.1.0
python-dateutil==2.4.2
pytz==2015.7
requests==2.9.1
six==1.10.0
